The pages have a shared navigation top bar
