+++ client/src/pages/dropdown/friends.tsx
@@ -1,10 +1,12 @@
 import React from "react";
+import Navbar from "../navbar";
 
 const Friends = () => {
   return (
     <div>
+      <Navbar />
       <h1>Friends Page</h1>
-      <p>Welcome to the Friends page!</p>
+      <p>Welcome to your friends settings!</p>
     </div>
   );
 };
