+++ client/src/styles/dropdown/settings.css
@@ -1,0 +1,270 @@
+/* Dashboard Container*/
+.dashboard-container {
+    padding: 60px 20px;
+    max-width: 800px;
+    margin: auto;
+    text-align: center;
+}
+
+/* Centering and spacing for Add Job Button */
+.add-job-container {
+    margin: 20px 0;
+}
+
+/* Add job button */
+.add-job-button {
+    background-color: #000;
+    color: #fff;
+    padding: 10px 16px;
+    border-radius: 6px;
+    cursor: pointer;
+    font-size: 16px;
+    border: none;
+    transition: background 0.2s ease-in-out;
+}
+
+/* Hover effect for add job button */
+.add-job-button:hover {
+    background-color: #333;
+}
+
+/* Job Form*/
+.job-form {
+    display: flex;
+    justify-content: center;
+    gap: 10px;
+    margin-bottom: 20px;
+    margin-top: 50px;
+    }
+
+.job-form input,
+.job-form select {
+    padding: 8px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+}
+
+.save-job-button {
+    background-color: #008000;
+    color: #fff;
+    padding: 8px 12px;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+.save-job-button:hover {
+    background-color: #006600;
+}
+
+/* Job Stats */
+.job-stats {
+    text-align: left;
+    background: #f8f9fa;
+    padding: 15px;
+    border-radius: 8px;
+    margin: 20px 0;
+}
+
+.job-stats ul {
+    list-style: none;
+    padding: 0;
+}
+
+.job-stats li {
+    padding: 4px 0;
+}
+
+/* Filter & Sorting */
+.filter-container {
+    display: flex;
+    justify-content: center;
+    gap: 10px;
+    margin-bottom: 20px;
+}
+
+.filter-button {
+    background: #e0e0e0;
+    padding: 8px 12px;
+    border-radius: 4px;
+    border: none;
+    cursor: pointer;
+}
+
+.filter-button:hover {
+    background: #d0d0d0;
+}
+
+.sort-dropdown {
+    padding: 8px;
+    border-radius: 4px;
+}
+
+/* Job List Section */
+.job-list-header {
+    margin-top: 30px;
+}
+
+.job-list {
+    padding: 20px;
+    border: 1px solid #ddd;
+    border-radius: 6px;
+    background: #fff;
+}
+
+.job-list p {
+    font-size: 14px;
+    color: #666;
+}
+
+/* The top navbar */
+.navbar {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 50px;  /* Adjust height as needed */
+    background-color: #ffffff;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0 20px;  /* Adds space on the sides */
+    border-bottom: 2px solid #ccc;
+    box-shadow: 0 2px 5px rgba(0,0,0,0.2);
+    z-index: 1000;
+}
+
+/* Left side (App title/logo) */
+.navbar-left {
+    font-size: 18px;
+    font-weight: bold;
+    color: #333;
+}
+
+/* Right side (user info + logout) */
+.navbar-right {
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    margin-right: 30px;
+}
+
+/* Profile picture container */
+.user-icon-container {
+    width: 32px;
+    height: 32px;
+    border-radius: 50%;
+    background-color: #ffffff;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border: 2px solid #ccc; 
+    padding: 2px; 
+    box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.2); 
+}
+
+/* Default user icon */
+.user-icon {
+    width: 100%; 
+    height: 100%;
+    border-radius: 50%;
+    object-fit: cover; 
+}
+
+/* Name with border */
+.name {
+    display: inline-block;
+    border: 1px solid #8a8a8a;
+    padding: 4px 10px;
+    border-radius: 4px;
+    font-size: 14px;
+}
+
+/* Logout button */
+.logout-button {
+    background-color: #000;
+    color: #fff;
+    padding: 6px 12px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    transition: background 0.2s ease-in-out;
+}
+
+/* Logout button hover effect */
+.logout-button:hover {
+    background-color: #333;
+}
+
+/* User Dropdown */
+.user-dropdown {
+    position: relative;
+    display: inline-block;
+    cursor: pointer;
+}
+
+/* Dropdown menu */
+.dropdown-menu {
+    position: absolute;
+    top: 100%;
+    right: 0;
+    background: white;
+    border: 1px solid #ddd;
+    border-radius: 6px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    padding: 8px;
+    min-width: 150px;
+    z-index: 100;
+}
+
+/* Dropdown items */
+.dropdown-item {
+    display: block;
+    width: 100%;
+    text-align: left;
+    padding: 10px;
+    border: none;
+    background: none;
+    font-size: 14px;
+    cursor: pointer;
+    transition: background 0.2s;
+}
+
+/* Dropdown items hover effect*/
+.dropdown-item:hover {
+    background: #f1f1f1;
+}
+
+.dropdown-menu hr {
+    border: 0.5px solid #ddd;
+    margin: 8px 0;
+}
+
+.dropdown-item.logout {
+    color: red;
+    font-weight: bold;
+}
+
+.dropdown-item.logout:hover {
+    background: #ffe6e6;
+}
+
+
+/* Main Logo button */
+.main-button {
+    display: block;
+    width: 100%;
+    text-align: left;
+    padding: 0; 
+    border: none;
+    background: none;
+    font-size: 32px; 
+    font-weight: bold;
+    color: inherit;
+    cursor: pointer;
+    transition: color 0.2s ease-in-out;
+}
+
+/* Optional: Add hover effect */
+.main-button:hover {
+    color: #555; /* Slightly darker on hover */
+}
